Switched taskMutated subscription call from Board component into BoardPage component, so that the listening client will get the newly added data even if is viewing the swimlane view
