+++ src/components/board/Board.jsx
@@ -1,44 +1,16 @@
 /* eslint-disable react/jsx-props-no-spreading */
 /* eslint-disable max-len */
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import { Grid } from '@material-ui/core'
 import { DragDropContext, Droppable } from 'react-beautiful-dnd'
-import { useApolloClient, useSubscription, gql } from '@apollo/client'
+import { useApolloClient } from '@apollo/client'
 import ColumnList from '../column/ColumnList'
 import useMoveTicketInColumn from '../../graphql/ticket/hooks/useMoveTicketInColumn'
 import useMoveTicketFromColumn from '../../graphql/ticket/hooks/useMoveTicketFromColumn'
 import useMoveColumn from '../../graphql/column/hooks/useMoveColumn'
 import { onDragEnd } from '../../utils/onDragEnd'
 import SnackbarAlert from '../SnackbarAlert'
-import { TICKETORDER_AND_TASKS, SWIMLANE_ORDER } from '../../graphql/fragments'
 import '../../styles.css'
-
-const TASK_SUBSCRIPTION = gql`
-  subscription taskMutated($boardId: ID!) {
-    taskMutated(boardId: $boardId) {
-      mutationType
-      node {
-        id
-        title
-        size
-        owner {
-          id
-          userName
-        }
-        members {
-            id
-            userName
-        }
-        description
-        prioritized
-        swimlaneOrderNumber
-        column {
-            id
-        }
-      }
-    }
-  }
-`
 
 const Board = ({ board }) => {
     const [moveTicketInColumn] = useMoveTicketInColumn()
@@ -47,32 +19,6 @@ const Board = ({ board }) => {
     const client = useApolloClient()
     const [snackbarStatus, setSnackbarStatus] = useState(false)
     const [snackbarMessage, setSnackbarMessage] = useState(null)
-    const { data, loading } = useSubscription(
-        TASK_SUBSCRIPTION,
-        { variables: { boardId: board.id } },
-
-    )
-    useEffect(() => {
-        if (!data) return
-        console.log('herre', data.taskMutated.node)
-        // Update the column's tasks and ticketOrder lists
-        const addedTask = data.taskMutated.node
-        const columnIdForCache = `Column:${addedTask.column.id}`
-        const { ticketOrder, tasks } = client.readFragment({
-            id: columnIdForCache,
-            fragment: TICKETORDER_AND_TASKS,
-        })
-        const newTasks = tasks.concat(addedTask)
-        const newTicketOrder = ticketOrder.concat({ ticketId: addedTask.id, type: 'task' })
-        client.writeFragment({
-            id: columnIdForCache,
-            fragment: TICKETORDER_AND_TASKS,
-            data: {
-                ticketOrder: newTicketOrder,
-                tasks: newTasks,
-            },
-        })
-    }, [data, client])
 
     const toggleSnackbar = (message) => {
         setSnackbarMessage(message)
